src/microservices/events/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
kafka-python==2.2.15